services/embed-worker/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1
aio-pika==9.5.5
asyncpg==0.30.0
aiofiles==24.1.0
pdfminer.six==20231228
python-docx==1.1.2
structlog==25.2.0
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-instrumentation-aio-pika==0.55b0
